requirements.txt
numpy==1.26.4
colour-demosaicing==0.2.5
colour-science==0.4.4
rawpy==0.23.2
opencv-contrib-python-headless==4.11.0.86
scikit-image==0.24.0
pillow==10.4.0
matplotlib==3.9.3
pytorch-msssim==1.0.0
h5py==3.13.0
tensorboard==2.18.0
tqdm==4.67.1
tabulate==0.9.0